raspberrypi/requirements.txt
Adafruit-GPIO==1.0.3
Adafruit-Nokia-LCD==0.2.0
Adafruit-PureIO==0.2.1
certifi==2018.4.16
chardet==3.0.4
future==0.16.0
idna==2.6
influxdb==5.0.0
iso8601==0.1.12
Pillow==5.1.0
pyserial==3.4
python-dateutil==2.7.2
pytz==2018.4
PyYAML==3.12
requests==2.18.4
six==1.11.0
urllib3==1.22
